requirements.txt
pdf2image
Pillow
python-dotenv
requests
argparse
pytest
setuptools